# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "jwt-service-app",
    "description": "",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "1.17.13"
  },
  "paths": {
...
    "/tokens/{jti}": {
      "delete": {
        "tags": [
          "handlers"
        ],
        "summary": "Revokes a token by its `jti` identifier.",
        "description": "It deletes the `jti` record from Redis; after that, verifying the\ncorresponding token in [`verify_token`] fails.\n\n# Responses\n- `204 No Content` — the token was revoked. **Idempotent**: a `jti` that does\n  not exist also gives `204`, because the desired state has been reached —\n  there is no such token;\n- `500 Internal Server Error` — the store is unavailable and the revocation\n  was **not** performed. Telling this case apart from success is mandatory:\n  the caller is revoking a compromised token and must learn that the attempt\n  failed.\n\nGeneric over the `jti` store ([`JtiStore`]).",
        "operationId": "revoke_token",
        "parameters": [
          {
            "name": "jti",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The token was revoked. Idempotent: a `jti` that does not exist also gives 204"
          },
          "401": {
            "description": "Level 3: the TOTP code is missing or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "The global cap of the endpoint was exceeded (when enabled)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "The store is unavailable — the revocation was NOT performed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "totp": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "description": "The unified body of an error response.",
        "required": [
          "error"
        ],
        "properties": {
          "details": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional details."
          },
          "error": {
            "type": "string",
            "description": "A short error message."
          }
        }
      },
...
    },
    "securitySchemes": {
      "metrics_token": {
        "type": "http",
        "scheme": "bearer",
        "description": "Level 4: a static bearer token for scraping /metrics (AUTH_METRICS_TOKEN)."
      },
      "proxy_secret": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Proxy-Secret",
        "description": "Level 2: a static secret injected by the reverse proxy. The proxy MUST strip the client-supplied version of the header."
      },
      "totp": {
        "type": "apiKey",
        "in": "header",
        "name": "X-TOTP-Code",
        "description": "Level 3: the current TOTP code (RFC 6238) over the shared secret."
      }
    }
  }
}